Reorganize asset models module
